api/gemini-proxy/requirements.txt
functions-framework==3.5.0
flask==3.0.0
google-cloud-aiplatform==1.38.0
